Move calculation of aggregated emissions data into emissions calculations file
