+++ client/src/app/components/MonthlyEmissionsCharts.tsx
@@ -6,10 +6,6 @@ import { useDispatch } from 'react-redux';
 import { ErrorBoundary } from 'app/components/ErrorBoundary';
 import { useTypedSelector } from 'app/redux/index';
 import type {
-  EmissionsData,
-  AggregatedEmissionsData,
-} from 'app/redux/reducers/results';
-import type {
   Aggregation,
   Source,
   Unit,
@@ -23,6 +19,10 @@ import {
   setMonthlyEmissionsSource,
   setMonthlyEmissionsUnit,
 } from 'app/redux/reducers/monthlyEmissions';
+import type {
+  EmissionsData,
+  AggregatedEmissionsData,
+} from 'app/calculations/emissions';
 import { useSelectedRegion, useSelectedStateRegions } from 'app/hooks';
 import type { Pollutant, RegionId, StateId } from 'app/config';
 import { regions, states } from 'app/config';
