+++ client/src/app/components/StateEmissionsTable.tsx
@@ -2,8 +2,8 @@ import { Fragment } from 'react';
 // ---
 import { ErrorBoundary } from 'app/components/ErrorBoundary';
 import { useTypedSelector } from 'app/redux/index';
-import type { AggregatedEmissionsData } from 'app/redux/reducers/results';
 import type { VehicleEmissionChangesByGeography } from 'app/calculations/transportation';
+import type { AggregatedEmissionsData } from 'app/calculations/emissions';
 import type { StateId } from 'app/config';
 import { states as statesConfig } from 'app/config';
